Important!!!!
update new work plan for different solutions. Please read it and feel free to modify it before we have a discussion tomorrow

I have already attach links for solution 2, this will help u guys evaluate the difficulty

Please help complete that for the solution 3.

The file is in schedule/Deng/