Render:
The library is basically finished, later we can do a little tweak to optimize it.(efficiency)

Interaction:

Drag and zoom is also implemented. Now we can also use mouse left click to select an item. This functions now can only display the id of the object. Later if we have time, we can add on a pop-out page to display the information

Todo:

1. Build more map for place searching
2. An interface for routing display
3. pop-out page for detail information
4. Release the Alpha version!!!!!!!!! Welcome to Le Creusot!!!!!!!!